Backwards compatibility for map configurations.
Mapping gxp_layertree to the LayerManager plugin, so old maps can take advantage of the LayerManager. Also switch "westpanel" id back to "tree", like it was in old versions. Updating gxp to be more forgiving when a container is missing (so we don't fail when trying to add the LegendPanel, which we don't have a container for any more).